# Improvement: Optimize LED library for maximum performance