### What
This PR aims to decrease CPU load and resource usage by optimizing the LED handler

### Why
Faster 10ms loop has nice impact on system overall CPU load

### How

Before optimization
Core task max load last 10 s: 216 us

After optimization;
Core task max load last 10 s: 206 us